quantum_vision/requirements.txt
pennylane>=0.36.0
torch>=2.2.0
torchvision>=0.17.0
opencv-python>=4.9.0
flask>=3.0.0
pyyaml>=6.0.0
matplotlib>=3.8.0
numpy>=1.26.0
